# docs/swagger.yaml
basePath: /
definitions:
  domain.FAQTableRecord:
    properties:
      is_resolved:
        type: string
      record:
        additionalProperties: {}
        type: object
      record_id:
        type: string
    type: object
  domain.TableRecord:
    properties:
      record:
        additionalProperties: {}
        type: object
      record_id:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  v1.CreatTableRecordReg:
    properties:
      contact_info:
        description: 联系方式，可选
        type: string
      content:
        description: 反馈内容
        type: string
      extra_record:
        additionalProperties: {}
        description: 额外记录列表，可选
        type: object
      images:
        description: 图片附件 URL 列表，可选
        items:
          type: string
        type: array
      table_identify:
        type: string
    required:
    - content
    - table_identify
    type: object
  v1.CreatTableRecordResp:
    properties:
      record_id:
        type: string
    type: object
...
  v1.FAQResolutionUpdateReq:
    properties:
      is_resolved:
        type: boolean
      record_id:
        type: string
      resolved_field_name:
        type: string
      table_identify:
        type: string
      unresolved_field_name:
        type: string
    required:
    - is_resolved
    - record_id
    - resolved_field_name
    - table_identify
    - unresolved_field_name
    type: object
...
  v1.GetTableRecordByRecordIdResp:
    properties:
      record:
        additionalProperties: {}
        type: object
    type: object
  v1.GetTableRecordResp:
    properties:
      has_more:
        type: boolean
      page_token:
        type: string
      records:
        items:
          $ref: '#/definitions/domain.TableRecord'
        type: array
      total:
        type: integer
    type: object
...
  v2.FAQResolutionUpdateReq:
    properties:
      is_resolved:
        type: boolean
      record_id:
        type: string
      table_identify:
        type: string
    required:
    - is_resolved
    - record_id
    - table_identify
    type: object
  v2.ForceSyncTableRecordsReq:
    properties:
      table_identify:
        type: string
    required:
    - table_identify
    type: object
  v2.ForceSyncTableRecordsResp:
    properties:
      queue_full:
        description: 队列是否已满
        type: boolean
      record_ids:
        description: 成功投递到队列的记录 ID
        items:
          type: string
        type: array
      total:
        description: 本次尝试同步的总记录数
        type: integer
    type: object
  v2.ForceSyncUserTableRecordsReq:
    properties:
      table_identify:
        type: string
    required:
    - table_identify
    type: object
  v2.ForceSyncUserTableRecordsResp:
    properties:
      queue_full:
        description: 队列是否已满
        type: boolean
      record_ids:
        description: 成功投递到队列的记录 ID
        items:
          type: string
        type: array
      total:
        description: 本次尝试同步的总记录数
        type: integer
    type: object
  v2.GetTableRecordByRecordIdResp:
    properties:
      records:
        items:
          $ref: '#/definitions/domain.FAQTableRecord'
        type: array
    type: object
  v2.GetTableRecordResp:
    properties:
      has_more:
        type: boolean
      page_token:
        type: string
      records:
        items:
          $ref: '#/definitions/domain.TableRecord'
        type: array
    type: object
  v2.SyncFaqRecordReq:
    properties:
      table_identify:
        type: string
    required:
    - table_identify
    type: object
  v2.SyncUnsyncedTableRecordsReq:
    properties:
      table_identify:
        type: string
    required:
    - table_identify
    type: object
  v2.SyncUnsyncedTableRecordsResp:
    properties:
      queue_full:
        description: 队列是否已满
        type: boolean
      record_ids:
        description: 成功投递到队列的记录 ID
        items:
          type: string
        type: array
      total:
        description: 本次尝试同步的总记录数
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  title: 木犀反馈系统 API
  version: "1.0"
paths:
...
  /api/v1/sheet/records:
    get:
      consumes:
      - application/json
      description: 根据当前 JWT 中的学生身份查询历史反馈记录，支持分页查询。前端不再传入学生 ID。
      operationId: get-table-record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用于查询记录的字段名，字段值由 JWT 中的学生身份决定
        in: query
        name: key_field
        required: true
        type: string
      - description: 分页参数,第一次不需要
        in: query
        name: page_token
        type: string
      - collectionFormat: csv
        description: 需要查询的字段名
        in: query
        items:
          type: string
        name: record_names
        required: true
        type: array
      - in: query
        name: table_identify
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回查询结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.GetTableRecordResp'
              type: object
        "400":
          description: 请求参数错误或飞书接口调用失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询历史反馈记录
      tags:
      - Sheet
    post:
      consumes:
      - application/json
      description: 向指定的多维表格应用中添加用户反馈记录，支持文本内容、截图附件和联系方式。创建成功后会异步发送通知消息。
      operationId: create-table-record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新增记录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreatTableRecordReg'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回创建记录结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.CreatTableRecordResp'
              type: object
        "400":
          description: 请求参数错误或飞书接口调用失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建反馈记录
      tags:
      - Sheet
  /api/v1/sheet/records/faq:
    get:
      consumes:
      - application/json
      description: 根据当前 JWT 中的学生身份查询相关的常见问题记录及其解决状态。
      operationId: get-faq-resolution-record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - collectionFormat: csv
        description: 需要查询的字段名
        in: query
        items:
          type: string
        name: record_names
        required: true
        type: array
      - in: query
        name: table_identify
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回单条记录的字段键值对
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.GetTableRecordByRecordIdResp'
              type: object
        "400":
          description: 请求参数错误或飞书接口调用失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询FAQ问题记录
      tags:
      - Sheet
    post:
      consumes:
      - application/json
      description: 用户更新FAQ问题的解决状态，将问题标记为已解决或未解决。
      operationId: update-faq-resolution
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新FAQ解决状态请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.FAQResolutionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新FAQ解决状态
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误或飞书接口调用失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 标记FAQ问题解决状态
      tags:
      - Sheet
  /api/v2/sheet/records:
    get:
      consumes:
      - application/json
      description: 根据当前 JWT 中的学生身份查询用户历史反馈记录，支持分页查询，前端不再传入学生 ID。
      operationId: get-table-record-by-user
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: limit_size
        type: integer
      - description: 分页参数,第一次不需要
        in: query
        name: page_token
        type: string
      - in: query
        name: table_identify
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回查询结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v2.GetTableRecordResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询用户历史反馈记录
      tags:
      - SheetV2
  /api/v2/sheet/records/faq:
    get:
      consumes:
      - application/json
      description: 根据当前 JWT 中的学生身份查询相关的常见问题记录及其解决状态。
      operationId: get-faq-record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: table_identify
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回单条记录的字段键值对
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v2.GetTableRecordByRecordIdResp'
              type: object
        "400":
          description: 请求参数错误或飞书接口调用失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询FAQ问题记录
      tags:
      - SheetV2
    post:
      consumes:
      - application/json
      description: 用户更新FAQ问题的解决状态，将问题标记为已解决或未解决。
      operationId: update-faq-resolution-v2
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新FAQ解决状态请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.FAQResolutionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新FAQ解决状态
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误或飞书接口调用失败
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 标记FAQ问题解决状态
      tags:
      - SheetV2
  /api/v2/sheet/sync:
    post:
      consumes:
      - application/json
      description: 同步指定表格下 is_synced = false 的所有记录，用于后台增量同步。
      operationId: sync-unsynced-table-records
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 同步请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.SyncUnsyncedTableRecordsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 同步成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v2.SyncUnsyncedTableRecordsResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 同步未同步记录
      tags:
      - SheetV2
  /api/v2/sheet/sync/faq:
    post:
      consumes:
      - application/json
      description: 同步某表格下的FAQ问题记录及其解决状态，用于后台增量同步或修复数据。
      operationId: sync-faq-record
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 同步FAQ记录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.SyncFaqRecordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功同步FAQ记录
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 同步FAQ问题记录
      tags:
      - SheetV2
  /api/v2/sheet/sync/force:
    post:
      consumes:
      - application/json
      description: 同步某表格下的所有记录（不区分是否已同步），用于全量重建或修复数据。慎用！！！
      operationId: force-sync-table-records
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 强制同步请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.ForceSyncTableRecordsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 同步成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v2.ForceSyncTableRecordsResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 强制同步表格所有记录
      tags:
      - SheetV2
  /api/v2/sheet/sync/user:
    post:
      consumes:
      - application/json
      description: 同步指定用户在某表格下的所有记录（不区分是否已同步），用于全量重建或修复数据。
      operationId: force-sync-user-table-records
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 强制同步请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.ForceSyncUserTableRecordsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 同步成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v2.ForceSyncUserTableRecordsResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 强制同步用户记录
      tags:
      - SheetV2
swagger: "2.0"